+++ app/src/androidTest/java/com/example/clubdeportivo/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.clinicaseprise
+package com.example.clubdeportivo
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.clinicaseprise", appContext.packageName)
+        assertEquals("com.example.clubdeportivo", appContext.packageName)
     }
 }
