+++ app/src/main/AndroidManifest.xml
@@ -10,8 +10,11 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.ClinicaSeprise"
+        android:theme="@style/Theme.ClubDeportivo"
         tools:targetApi="31">
+        <activity
+            android:name=".IngresarEmpleados"
+            android:exported="false" />
         <activity
             android:name=".MainActivity"
             android:exported="true">
