+++ app/src/main/java/com/example/clubdeportivo/ErrorCredenciales.kt
@@ -1,4 +1,4 @@
-package com.example.clinicaseprise
+package com.example.clubdeportivo
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
@@ -6,11 +6,11 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 
-class ConexionExitosa : AppCompatActivity() {
+class ErrorCredenciales : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_conexion_exitosa)
+        setContentView(R.layout.activity_error_credenciales)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
