+++ app/src/main/res/layout/activity_main.xml
@@ -1,69 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:layout_gravity="center"
-    android:padding="40dp"
-    tools:context=".MainActivity"
-    android:background="@drawable/gradient_background">
+    tools:context="com.example.clubdeportivo.MainActivity">
 
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Login"
-        android:textSize="24sp"
-        android:layout_gravity="center"
-        android:paddingBottom="32dp"/>
-
-    <ImageView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/character_image"
-        android:layout_gravity="center"
-        android:paddingBottom="16dp"/>
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/una_breve_descripci_n_sobre_la_nueva_caracter_stica_de_la_app"
-        android:layout_gravity="center"
-        android:paddingBottom="32dp"/>
-
-    <EditText
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:hint="Usuario"
-        android:padding="12dp"
-        android:background="@drawable/colors"
-        android:layout_marginBottom="16dp"/>
-
-    <EditText
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:hint="Contraseña"
-        android:inputType="textPassword"
-        android:padding="12dp"
-        android:background="@drawable/edit_text_background"
-        android:layout_marginBottom="32dp"/>
-
-    <Button
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Iniciar sesión"
-        android:padding="12dp"
-        android:background="#FF03DAC5"
-        android:textColor="@android:color/white"
-        android:layout_marginBottom="16dp"/>
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Antes"
-        android:layout_gravity="center"
-        android:textColor="#FF03DAC5"/>
-
-</LinearLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
