+++ app/src/test/java/com/example/clubdeportivo/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.clinicaseprise
+package com.example.clubdeportivo
 
 import org.junit.Test
